Implemented variable permission system
We needed a third level of permissions so it was high time to fix
permissions once and for all. It's implemented to bitwise operations
on unsigned 64-bit ints each bit being a particular permission. There
are currently 3 predefined permission sets or groups. Long term this
could be implemented in such a way as to be configurable by the end
user. For now it makes it a lot easier to add a new permission level
if needed.

All calls to IsAdmin, IsHelpDesk, and IsNormal have been replaced
with a call to Can() and checking for a particular permission vs
group membership.

Other changes in this commit:

- Bug fix for displaying Daily @ time on admin user edit page
- Added Copyright section to contributing guidelines
- Added device last_seen column to device manage page
- Created new readonly user (pervious help desk user) and gave the
help desk user more permissions
- Formatted validity dates on user list page
- Removed Onesimus Systems copyright statements and transitioned to
University of Southern Indiana
- Permission checks have been delegated to the handlers instead of
middleware due to the myriad of permissions

References #6

+++ src/controllers/api/device.go
@@ -36,15 +36,20 @@ func (d *Device) RegistrationHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if sessionUser.Username != formUsername && !sessionUser.IsAdmin() && !sessionUser.IsHelpDesk() {
+	if sessionUser.Username != formUsername && !sessionUser.Can(models.CreateDevice) {
 		d.e.Log.Errorf("Admin action attempted: Register device for %s attempted by user %s", formUsername, sessionUser.Username)
 		common.NewAPIResponse("Only admins can do that", nil).WriteResponse(w, http.StatusForbidden)
 		return
 	}
 
-	if formUsername == sessionUser.Username {
+	if formUsername == sessionUser.Username && sessionUser.Can(models.CreateOwn) {
 		formUser = sessionUser
 	} else {
+		common.NewAPIResponse("Permission denied", nil).WriteResponse(w, http.StatusForbidden)
+		return
+	}
+
+	if formUsername != sessionUser.Username {
 		var err error
 		formUser, err = models.GetUserByUsername(d.e, formUsername)
 		if err != nil {
@@ -54,7 +59,7 @@ func (d *Device) RegistrationHandler(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	if !sessionUser.IsAdmin() && !sessionUser.IsHelpDesk() {
+	if !sessionUser.Can(models.CreateDevice) {
 		if formUser.IsBlacklisted() {
 			d.e.Log.Noticef("Attempted registration by blacklisted user %s", formUser.Username)
 			common.NewAPIResponse("Username blacklisted", nil).WriteResponse(w, http.StatusForbidden)
@@ -155,7 +160,7 @@ func (d *Device) RegistrationHandler(w http.ResponseWriter, r *http.Request) {
 
 	// Redirect client as needed
 	resp := struct{ Location string }{Location: "/manage"}
-	if !sessionUser.IsNormal() {
+	if sessionUser.Can(models.ViewDevices) {
 		resp.Location = "/admin/manage/" + formUser.Username
 	}
 
@@ -171,7 +176,7 @@ func (d *Device) DeleteHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	if username != sessionUser.Username {
-		if !sessionUser.IsAdmin() {
+		if !sessionUser.Can(models.DeleteDevice) {
 			common.NewAPIResponse("Admin Error", nil).WriteResponse(w, http.StatusForbidden)
 			return
 		}
@@ -182,6 +187,11 @@ func (d *Device) DeleteHandler(w http.ResponseWriter, r *http.Request) {
 			common.NewAPIResponse("Server error", nil).WriteResponse(w, http.StatusInternalServerError)
 			return
 		}
+	}
+
+	if !sessionUser.Can(models.DeleteOwn) {
+		common.NewAPIResponse("Permission denied", nil).WriteResponse(w, http.StatusForbidden)
+		return
 	}
 
 	deleteAll := (r.FormValue("mac") == "")
@@ -217,8 +227,8 @@ func (d *Device) DeleteHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 func (d *Device) ReassignHandler(w http.ResponseWriter, r *http.Request) {
-	sessUser := models.GetUserFromContext(r)
-	if !sessUser.IsAdmin() {
+	sessionUser := models.GetUserFromContext(r)
+	if !sessionUser.Can(models.ReassignDevice) {
 		common.NewEmptyAPIResponse().WriteResponse(w, http.StatusForbidden)
 		return
 	}
@@ -269,7 +279,7 @@ func (d *Device) ReassignHandler(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 		d.e.Log.WithFields(verbose.Fields{
-			"adminUser":    sessUser.Username,
+			"adminUser":    sessionUser.Username,
 			"assignedTo":   username,
 			"assignedFrom": originalUser,
 			"MAC":          mac.String(),
